Day 10: Binary Numbers
Task 
Given a base-10 integer,n , convert it to binary (base-2). Then find and print the base-10 integer denoting the maximum number of consecutive 1's in n's binary representation.

Input Format
A single integer, n.

Output Format
Print a single base-10 integer denoting the maximum number of consecutive 1's in the binary representation of n.

Sample Case 1: 
The binary representation of 5 is 101, so the maximum number of consecutive 1's is 1.

Sample Case 2: 
The binary representation of 13 is 1101, so the maximum number of consecutive 1's is 2.